backend/NIKONEKTI_backend/requirements.txt
Django>=4.0
djangorestframework
drf-yasg
django-extensions
djangorestframework-authtoken
corsheaders
